requirements.txt
django
django-import-export
django-htmx
requests